maintain: move the human formatter out to its own package
