+++ internal/format/format.go
@@ -1,4 +1,4 @@
-package cmd
+package format
 
 import (
 	"fmt"
